I. #32 first attempt at a fix for layering bug
This bug was brought about after testing sorting but wasn't necessarily
caused by sorting. Basically the bug was not being able to reliably
redraw the z-index order when changing scale, changing basemaps, and
changing visibility. This commit is a first pass at fixing that. It
currently seems to function as expected.