I. #87 rework map clicks to allow protected areas
This commit is a pretty major rework of handling map clicks. There is
now one click event handler that depending on the current selected scale
will query features from stats-gadm0, stats-gadm1, stats-hybas. It will
then check for the only other clickable layer which is protected areas.
It will build up a html string response to put in a popup.

Also had to add duplicate layers for protected areas, one for showing
only the outline of the areas and another transparent fill to allow to
grab clicks.