I. #10 enhance mapbox draw control
Updated turf package. Using turf to get bbox of selected drawn features
to help offset the popup. This is so the popup does not cover the
highlighted area.

Set popup to default to show on the right side with a buffer.

Moved the custom draw style to its own file and importing.

Using a hack / non-reactjs way to update the tooltip text and color
styling for the mapbox draw buttons.