I. #28 disable search bars unless selected
Added some CSS trickery to make it appear as though the geocoder search
bars are disabled unless selected. In reality there are two components
occupying the same space. One, the mapbox geocoder, which is hidden and
not shown unless selected. The second, a react bootstrap input to look
like the mapbox search that is always disabled and visible when NOT
selected.

Also setting admin geocoder to use local gadm1 only to avoid duplicates
in search suggestions.