I. #32 reset eye vis when changing scales.
When changing scales we're really changing the data in most cases as
each aggegrated layer is it's own thing. Thus when using the eye icon to
turn on/off visibility, it doesn't translate to the layers at different
scales even though they might all be the same service.

When changing scales, make sure to reset the eye icons to visible.