I. #132 screen service layers before showing stats
Fix a bug where some layers like population, lulc, and others were not
being filtered out when building the stats html to show in a popup.
coastalProtection was also not handled for showing stats when drawing in
local mode, which caused a bug where a user could not double click to
close the polygon draw.